+++ Sources/StreamVideo/WebRTC/PeerConnectionFactory.swift
@@ -43,6 +43,14 @@ final class PeerConnectionFactory: @unchecked Sendable {
         defaultDecoder.supportedCodecs()
     }
 
+    func setPreferredEncodingCodec(_ videoCodec: VideoCodec) {
+        if let preferredCodec = defaultEncoder.supportedCodecs().first(where: { $0.name.lowercased() == videoCodec.rawValue }) {
+            defaultEncoder.preferredCodec = preferredCodec
+        } else {
+            log.warning("Unable to set preferred encoding codec \(videoCodec.rawValue).")
+        }
+    }
+
     /// Creates or retrieves a PeerConnectionFactory instance for a given
     /// audio processing module.
     /// - Parameter audioProcessingModule: The RTCAudioProcessingModule to use.
@@ -75,14 +83,30 @@ final class PeerConnectionFactory: @unchecked Sendable {
     /// - Parameter forScreenShare: Boolean indicating if the source is for screen sharing.
     /// - Returns: An RTCVideoSource instance.
     func makeVideoSource(forScreenShare: Bool) -> RTCVideoSource {
-        factory.videoSource(forScreenCast: forScreenShare)
+        let result = factory.videoSource(forScreenCast: forScreenShare)
+        log.debug(
+            """
+            VideoSource was created \(Unmanaged.passUnretained(result).toOpaque())
+            Encoder preferredCodec: \(defaultEncoder.preferredCodec)
+            """,
+            subsystems: .webRTC
+        )
+        return result
     }
 
     /// Creates a video track from a given video source.
     /// - Parameter source: The RTCVideoSource to use for the track.
     /// - Returns: An RTCVideoTrack instance.
     func makeVideoTrack(source: RTCVideoSource) -> RTCVideoTrack {
-        factory.videoTrack(with: source, trackId: UUID().uuidString)
+        let result = factory.videoTrack(with: source, trackId: UUID().uuidString)
+        log.debug(
+            """
+            VideoTrack was created \(Unmanaged.passUnretained(result).toOpaque())
+            trackId: \(result.trackId)
+            """,
+            subsystems: .webRTC
+        )
+        return result
     }
 
     /// Creates an audio source with optional constraints.