+++ StreamVideoTests/WebRTC/v2/PeerConnection/MediaAdapters/LocalMediaAdapters/LocalScreenShareMediaAdapter_Tests.swift
@@ -194,7 +194,7 @@ final class LocalScreenShareMediaAdapter_Tests: XCTestCase {
     func test_publish_disabledLocalTrack_transceiverHasBeenCreated_enablesAndAddsTrack() async throws {
         mockPeerConnection.stub(
             for: .addTransceiver,
-            with: try makeTransceiver(of: .screenshare, codecs: [VideoCodec.screenshare])
+            with: try makeTransceiver(of: .screenshare, layers: [VideoLayer.screenshare])
         )
         try await subject.beginScreenSharing(
             of: .inApp,
@@ -203,11 +203,11 @@ final class LocalScreenShareMediaAdapter_Tests: XCTestCase {
 
         let transceiver = try XCTUnwrap(mockPeerConnection.stubbedFunction[.addTransceiver] as? RTCRtpTransceiver)
         XCTAssertTrue(subject.localTrack?.isEnabled ?? false)
-        XCTAssertEqual(mockPeerConnection.stubbedFunctionInput[.addTransceiver]?.count, 1)
+        XCTAssertEqual(mockPeerConnection.timesCalled(.addTransceiver), 1)
         /// When there is only one encoding on the transceiver, WebRTC internally, removes the rid
         XCTAssertEqual(
             transceiver.sender.parameters.encodings.compactMap(\.maxBitrateBps),
-            [NSNumber(value: VideoCodec.screenshare.maxBitrate)]
+            [NSNumber(value: VideoLayer.screenshare.maxBitrate)]
         )
     }
 
@@ -235,7 +235,7 @@ final class LocalScreenShareMediaAdapter_Tests: XCTestCase {
         of type: TrackType,
         direction: RTCRtpTransceiverDirection = .sendOnly,
         streamIds: [String] = [.unique],
-        codecs: [VideoCodec]? = nil
+        layers: [VideoLayer]? = nil
     ) throws -> RTCRtpTransceiver {
         if temporaryPeerConnection == nil {
             temporaryPeerConnection = try peerConnectionFactory.makePeerConnection(
@@ -251,7 +251,7 @@ final class LocalScreenShareMediaAdapter_Tests: XCTestCase {
                 trackType: type,
                 direction: direction,
                 streamIds: streamIds,
-                codecs: codecs
+                layers: layers
             )
         )!
     }