+++ StreamVideoTests/WebRTC/v2/PeerConnection/MediaAdapters/VideoCapturePolicy/AdaptiveVideoCapturePolicy_tests.swift
@@ -729,7 +729,7 @@ final class AdaptiveVideoCapturePolicy_Tests: XCTestCase {
     // MARK: - Private helpers
 
     private func assertUpdateCaptureQuality(
-        expected: [VideoCodec],
+        expected: [VideoLayer],
         expectedTimesCalled: Int,
         neuralEngineExists: Bool,
         thermalState: ProcessInfo.ThermalState,
@@ -743,7 +743,7 @@ final class AdaptiveVideoCapturePolicy_Tests: XCTestCase {
 
         try await subject.updateCaptureQuality(
             with: .init(
-                expected.map(\.quality)
+                expected.map(\.quality.rawValue)
             ),
             for: activeCaptureSession
         )
@@ -757,10 +757,10 @@ final class AdaptiveVideoCapturePolicy_Tests: XCTestCase {
 
         if expectedTimesCalled > 0 {
             XCTAssertEqual(
-                cameraVideoCapturer.recordedInputPayload(([VideoCodec], AVCaptureDevice?).self, for: .updateCaptureQuality)?.first?
+                cameraVideoCapturer.recordedInputPayload(([VideoLayer], AVCaptureDevice?).self, for: .updateCaptureQuality)?.first?
                     .0
-                    .map(\.quality).sorted(),
-                expected.map(\.quality).sorted(),
+                    .map(\.quality.rawValue).sorted(),
+                expected.map(\.quality.rawValue).sorted(),
                 file: file,
                 line: line
             )